Move PaymentChannelStateService into mpe package
