+++ sdk/src/main/java/io/singularitynet/sdk/client/Sdk.java
@@ -16,12 +16,12 @@ import io.singularitynet.sdk.registry.CachingMetadataProvider;
 import io.singularitynet.sdk.registry.RegistryMetadataProvider;
 import io.singularitynet.sdk.daemon.DaemonConnection;
 import io.singularitynet.sdk.daemon.RandomEndpointDaemonConnection;
-import io.singularitynet.sdk.daemon.PaymentChannelStateService;
 import io.singularitynet.sdk.mpe.AskDaemonFirstPaymentChannelProvider;
-import io.singularitynet.sdk.mpe.MultiPartyEscrowContract;
-import io.singularitynet.sdk.mpe.PaymentChannelStateProvider;
 import io.singularitynet.sdk.mpe.BlockchainPaymentChannelManager;
 import io.singularitynet.sdk.mpe.MpePaymentChannelManager;
+import io.singularitynet.sdk.mpe.MultiPartyEscrowContract;
+import io.singularitynet.sdk.mpe.PaymentChannelStateProvider;
+import io.singularitynet.sdk.mpe.PaymentChannelStateService;
 import io.singularitynet.sdk.freecall.FreeCallStateService;
 import io.singularitynet.sdk.client.PaymentStrategy;
 import io.singularitynet.sdk.client.ServiceClient;
